requirements.txt
av==9.2.0
boto3==1.34.87
botocore==1.34.87
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
jmespath==1.0.1
pillow==10.3.0
python-dateutil==2.9.0.post0
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
urllib3==1.26.18
Flask==2.0.3